Split CMakeLists files and remove unused deps and old ADM files (#66)
* Split CMakeLists files. Removed unused deps and old ADM files.

* Fixing coverage generation for separate target libraries

* Reporting coverage text for jobs

+++ deps.sh
@@ -30,7 +30,7 @@ BUILDDIR=${BUILDDIR:-${SELFDIR}/deps/build}
 echo "Building in ${BUILDDIR}"
 echo "Installing to ${DESTDIR}"
 
-if [ ! -e ${DESTDIR}/usr/include/ion.h ]
+if [ false && ! -e ${DESTDIR}/usr/include/ion.h ]
 then
   mkdir -p ${BUILDDIR}
   rsync --recursive ${DEPSDIR}/ion/ ${BUILDDIR}/ion/
